🥂 adds custom Toast component, /auth route using react router's useLocation
