+++ src/components/ProtectedRoute.jsx
@@ -1,14 +1,20 @@
-import { useSelector } from "react-redux";
-import { Outlet } from "react-router-dom";
-import { Text } from "@chakra-ui/react";
-import SignUp from "./auth/SignUp";
+import { useEffect } from "react";
+import { useDispatch, useSelector } from "react-redux";
+import { Outlet, useLocation, useNavigate } from "react-router-dom";
+import { setMessage } from "../state/features/messageSlice";
 
 const ProtectedRoute = () => {
+  const location = useLocation();
   const { currentUser } = useSelector((state) => state.user);
+  const dispatch = useDispatch();
+  const navigate = useNavigate();
 
-  if (!currentUser) {
-    return <SignUp message="You can't do that!" />;
-  }
+  useEffect(() => {
+    if (!currentUser) {
+      dispatch(setMessage(["You can't do that!"]));
+      navigate("/auth", { state: { originalRoute: location } });
+    }
+  }, [currentUser, location]);
 
   return <Outlet />;
 };
